third_party/blink/web_tests/platform/generic/http/tests/devtools/console/viewport-testing/console-clicking-messages-expected.txt

Tests that console messages are navigable with the keyboard.


Running: testClickOnLog
Evaluating: console.log(1)
Message count: 1
Click on message
Viewport virtual selection: 0
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected

Running: testClickOnGroup
Evaluating: console.group('group1')
Message count: 1
Click on message
Viewport virtual selection: 0
Is group expanded: NO
activeElement: DIV.console-message

Running: testClickOnTrace
Evaluating: console.warn('warning1')
Message count: 1
Click on message
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: DIV

Running: testClickOnObject
Evaluating: console.log({x: 1})
Message count: 1
Click on object
Viewport virtual selection: 0
Has object: expanded
activeElement: SPAN.console-object.console-object-preview

Running: testClickOnTraceWithObject
Evaluating: console.warn('warn', {x: 1})
Message count: 1
Click on object
Viewport virtual selection: 0
Has object: expanded
Is trace expanded: NO
activeElement: SPAN.console-object.console-object-preview
Click on trace
Viewport virtual selection: 0
Has object: expanded
Is trace expanded: YES
activeElement: DIV

Running: testClickOnGroupWithObject
Evaluating: console.group('group', {x: 1})
Message count: 1
Click on object
Viewport virtual selection: 0
Has object: expanded
Is group expanded: YES
activeElement: SPAN.console-object.console-object-preview
Click on group
Viewport virtual selection: 0
Has object: expanded
Is group expanded: NO
activeElement: DIV.console-message
